gpa: Fix the types in _add_cyclic.
- `int64_t` -> `int32_t`
- `cycles` could potentially be negative too.

+++ libhpx/gas/gpa.c
@@ -107,7 +107,7 @@ static hpx_addr_t _add_cyclic(hpx_addr_t gpa, int64_t n, uint32_t bsize,
 
   uint32_t _phase = _phase_of(gpa, bsize) + n;
   uint32_t phase  = _phase % bsize;
-  int64_t  blocks = _phase / bsize;
+  int32_t  blocks = _phase / bsize;
   if (_phase < 0) {
     phase += bsize;
     blocks = (_phase + 1) / bsize - 1;
@@ -115,7 +115,7 @@ static hpx_addr_t _add_cyclic(hpx_addr_t gpa, int64_t n, uint32_t bsize,
 
   uint32_t _rank  = gpa_to_rank(gpa) + blocks;
   uint32_t rank   = _rank % here->ranks;
-  uint32_t cycles = _rank / here->ranks;
+  int32_t cycles = _rank / here->ranks;
   if (_rank < 0) {
     rank  += here->ranks;
     cycles = (_rank + 1) / here->ranks - 1;
